Support multiple sets of icons
This adds a further level of indirection to the mapping of icons to
presets which allows preset files with references to completely
different icons to be generated.

The preset file now contains a simple tag that will be replaced by the
actual icon path or in the case of "beautified" icon set a template that
in turn is replaced during the build process.